[fix] correctly reset context upon corrupted RLE packet
When a wrong seq no or CRC is detected, the context is flushed then
reset in greater depth. The reset overwrites a part of the flushed
context by mistake: the end_address pointer used on the ZC_buffer
is set to NULL during reset, thus erasing the value set by the flush.
Because of that NULL pointer, the RLE library segfaults when handling
the next RLE packet.

The order of flush/reset was changed to avoid the NULL overwrite and
keep the expected value.