Adding a test to validate the context flushing.
Upon corrupted RLE packet, the context is flushed and reseted. this
flushing leaded to a segfault until the fix, however, there is no PCAP
trace that could lead to this segfault as the test_non_regression_fpdu
stop at the wrong seq no. This test allows checking that after a
corrupted packet, the flushing and reseting of the context does not
lead to the segfault.