Design modification to handle RLE header corruption on frag failure.
During RLE fragmentation, an event can occure :
The library caller wants a PPDU from a given fragment id ;
The library caller's burst is small, but not enough to know that it is
possible or not to extract a PPDU (ie. headers cannot be fragmented, but
their sizes are not known before calling for a fragmentation) ;
The RLE library tries to extract a PPDU. The SDU has just been
encapsulated so the RLE library make a START PPDU, the RLE header from
the ZC buffer is modified ;
The RLE library realizes during the fragmentation that the buffer is to
small for the PPDU header ;
The fragmentation stops, and the RLE library return an error code to the
library caller, to inform him that the fragmentation failed due to a
too small buffer ;
The library caller is expected to ask for a bigger PPDU (for instance,
to populate the next FPDU) ;
The error occured because the RLE header in the ZC buffer were
compromised in the first call, and thus, the PPDU extracted on the
second call is wrong. Nevertheless this situation occured only on a
COMPLETE PPDU (the RLE header being considered as good in the context) ;
The Fix consists in modifiying the library to allow the fragmentation to
call a new encapsulation on fragmentation error, to regenerate the RLE
header in the ZC buffer.