Adding a performances test on FPDUs.
Adding a test that sniff FPDUs on IP from an interface, and decapsulate
them in SDUs for performances tests.
Caution: for this test, every packet sniffed on the interface is
considered as FPDUs, thus false negtives can exists. Moreover the size
of the Payload label is very important for this test and must absolutly
be the same as the one used for packing.