install includes in "librle" subddir
This allow two avoid conflict if user land librle is installed
on system and we try to build librle's kernel module.
If rle.h is installed in /usr/include/ the kernel module's Makefile
takes installed rle.h in /usr/include instead of it's local one in
source tree and it might be not compatible.