Revert "fix pkg_config"
This reverts commit 32c49fc809c3bfbf7161e618332bdabff2b722a2.

The public header rle.h was moved into a librle/ subdir in order to
avoid problems of build for the kernel module if the userspace librle
is installed on the system.

There is no need to install the librle.so library on the dev machines.
In dev mode deps/ are used. In release mode, RPM are used. In addition
the librle/ subdir breaks the RLD build.