receiver: synchronize on first received seqnum
The RLE receiver expected to receive 0 as first seqnum for all
fragmented frames. This does not stand in the real world: the RLE
transmitter is not restarted or reset at the very same time the RLE
receiver is.

So, the RLE receiver must accept any seqnum as the first seqnum for
every reassembly contexts. The RLE receiver must synchronize on those
seqnums, in order to expect seqnum + 1 for the next fragmented frame.